+++ src/index.ts
@@ -8,3 +8,4 @@ export * from "./mount.js";
 export * from "./focus.js";
 export * from "./aria.js";
 export * from "./roving-focus.js";
+export * from "./menu.js";